moved the favorites reducer into the reducers folder
