+++ app/reducers/favoritesReducer.js
