+++ app/reducers/index.js
@@ -1,12 +1,12 @@
-import { combineReducers } from "redux";
-import loggedIn from "./loginReducer";
-import favorites from "../components/Favorites/favoritesReducer";
-import filmsNowPlaying from "./filmsNowPlaying";
+import { combineReducers } from 'redux'
+import loggedIn from './loginReducer'
+import favorites from './favoritesReducer'
+import filmsNowPlaying from './filmsNowPlaying'
 
 const rootReducer = combineReducers({
-  filmsNowPlaying,
-  loggedIn,
-  favorites
-});
+	filmsNowPlaying,
+	loggedIn,
+	favorites
+})
 
-export default rootReducer;
+export default rootReducer
